eye-disease-detection/requirements.txt
Flask
Pillow
tensorflow
numpy
flask-cors
tensorflowjs